deletion only when not in use

+++ src/main/java/de/inhorn/cybhorn/controller/SubscriptionController.java
@@ -3,6 +3,7 @@ package de.inhorn.cybhorn.controller;
 import de.inhorn.cybhorn.assembler.SubscriptionAssembler;
 import de.inhorn.cybhorn.model.dtos.SubscriptionDto;
 import de.inhorn.cybhorn.repository.SubscriptionRepository;
+import de.inhorn.cybhorn.repository.SubscriberRepository;
 import de.inhorn.cybhorn.service.SubscriptionService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,6 +24,7 @@ import java.util.List;
 public class SubscriptionController {
 	private final SubscriptionRepository subscriptionRepository;
 	private final SubscriptionService subscriptionService;
+	private final SubscriberRepository subscriberRepository;
 
 	@GetMapping("/list")
 	public List<SubscriptionDto> getAllSubscribers() {
@@ -36,7 +38,13 @@ public class SubscriptionController {
 
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> deleteSubscription(@PathVariable long id) {
-		subscriptionRepository.deleteById(id);
-		return new ResponseEntity<>(HttpStatus.OK);
+		if(!subscriberRepository.existsBySubscription(subscriptionService.findById(id))){
+
+			subscriptionRepository.deleteById(id);
+
+			return new ResponseEntity<>(HttpStatus.OK);
+		} else {
+			throw new IllegalArgumentException("Terminal is still in use by subscribers");
+		}
 	}
 }
